+++ src/compas_timber/fabrication/__init__.py
@@ -2,6 +2,7 @@ from .btlx import BTLx
 from .btlx import BTLxProcess
 from .btlx_processes.btlx_french_ridge_lap import BTLxFrenchRidgeLap
 from .btlx_processes.btlx_jack_cut import BTLxJackCut
+from .btlx_processes.lap import BTLxLap
 from .joint_factories.french_ridge_factory import FrenchRidgeFactory
 from .joint_factories.l_butt_factory import LButtFactory
 from .joint_factories.l_miter_factory import LMiterFactory
@@ -11,6 +12,7 @@ __all__ = [
     "BTLx",
     "BTLxProcess",
     "BTLxJackCut",
+    "BTLxLap",
     "BTLxFrenchRidgeLap",
     "LButtFactory",
     "TButtFactory",